pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "peakrdl-beam"
dynamic = ["version"]
requires-python = ">=3.6"
dependencies = [
    "systemrdl-compiler >= 1.21.0, < 2",
    "jinja2",
]

authors = [
    {name="João Henrique Ferreira de Freitas"}
]
description = "Generate Erlang or Elixir modules from a SystemRDL register model"
readme = "README.md"
license = {file = "LICENSE"}
keywords = [
    "SystemRDL", "PeakRDL", "CSR", "compiler", "tool", "registers", "generator",
    "Erlang", "Elixir", "BEAM", "software",
]
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.6",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3 :: Only",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Operating System :: OS Independent",
    "Topic :: Scientific/Engineering :: Electronic Design Automation (EDA)",
]

[project.urls]
Source = "https://github.com/joaohf/PeakRDL-beam"
Tracker = "https://github.com/joaohf/PeakRDL-beam/issues"
Changelog = "https://github.com/joaohf/PeakRDL-beam/releases"
Documentation = "https://peakrdl-beam.readthedocs.io/"

[tool.setuptools.dynamic]
version = {attr = "peakrdl_beam.__about__.__version__"}

[project.entry-points."peakrdl.exporters"]
beam = "peakrdl_beam.__peakrdl__:Exporter"
